+++ gulpfile.js
@@ -65,6 +65,8 @@ function tineMCE(done) {
 // Gulp Frontend Script
 function frontScripts(done) {
     const jsFiles = [
+        './assets/dev/javascript/user-tracker.js',
+        './assets/dev/javascript/event-tracker.js',
         './assets/dev/javascript/tracker.js',
     ];
 
@@ -74,6 +76,7 @@ function frontScripts(done) {
         .pipe(replace("\\t", ''))
         .pipe(replace("  ", ''))
         .pipe(uglify())
+        .pipe(concat('tracker.js'))
         .pipe(gulp.dest('./assets/js/'));
 
     done()
@@ -144,6 +147,8 @@ function addEsnextSuffix(filePath) {
 
 function revertToES6(cb) {
     const jsFiles = [
+        './assets/js/user-tracker.js',
+        './assets/js/event-tracker.js',
         './assets/js/tracker.js',
     ];