+++ includes/admin/templates/layout/tabbed-page-header.php
@@ -1,5 +1,10 @@
 <?php
+
 use WP_STATISTICS\Helper;
+use WP_Statistics\Components\View;
+
+View::load('components/objects/share-anonymous-notice');
+
 ?>
 <div class="wps-wrap__top tabbed_page">
     <h2 class="wps_title">
@@ -16,12 +21,12 @@ use WP_STATISTICS\Helper;
     <?php if (isset($real_time_button)): ?>
         <?php
         $is_realtime_active = Helper::isAddOnActive('realtime-stats');
-          if ($is_realtime_active): ?>
+        if ($is_realtime_active): ?>
             <a class="wps-realtime-btn" href="<?php echo esc_url(admin_url('admin.php?page=wp_statistics_realtime_stats')) ?>" title="<?php echo esc_html_e('Real-time stats are available! Click here to view', 'wp-statistics') ?>">
                 <?php esc_html_e('Realtime', 'wp-statistics'); ?>
             </a>
         <?php else: ?>
-            <button class="wps-realtime-btn disabled wps-tooltip-premium" >
+            <button class="wps-realtime-btn disabled wps-tooltip-premium">
                 <?php esc_html_e('Realtime', 'wp-statistics'); ?>
                 <span class="wps-tooltip_templates tooltip-premium tooltip-premium--bottom tooltip-premium--right">
                     <span id="tooltip_realtime">
@@ -34,7 +39,6 @@ use WP_STATISTICS\Helper;
             </button>
         <?php endif ?>
     <?php endif; ?>
-
 
 
     <div class="wp-clearfix"></div>
@@ -76,7 +80,7 @@ use WP_STATISTICS\Helper;
                     <?php if (isset($tab['coming_soon'])): ?>
                         <span class="wps-tooltip wps-tooltip--coming_soon" title="<?php echo esc_html__('Coming soon', 'wp-statistics') ?>"><?php echo esc_html($tab['title']); ?> <i class="wps-tooltip-icon coming-soon"></i></span>
                     <?php elseif (isset($tab['locked'])) : ?>
-                        <a  data-target="wp-statistics-data-plus"  class="js-wps-openPremiumModal wps-locked">
+                        <a data-target="wp-statistics-data-plus" class="js-wps-openPremiumModal wps-locked">
                             <?php echo esc_html($tab['title']); ?>
                             <?php if (!empty($tab['tooltip'])) : ?>
                                 <span class="wps-tooltip" title="<?php echo esc_attr($tab['tooltip']) ?>"><i class="wps-tooltip-icon info"></i></span>