+++ includes/admin/templates/layout/title.php
@@ -6,6 +6,7 @@ use WP_STATISTICS\Admin_Template;
 use WP_Statistics\Utils\Request;
 use WP_Statistics\Components\View;
 
+View::load('components/objects/share-anonymous-notice');
 ?>
 <div class="wps-wrap__top <?php echo isset($real_time_button) ? 'wps-wrap__top--has__realtime' : ''; ?>">
     <?php if (isset($backUrl, $backTitle)): ?>
@@ -29,7 +30,6 @@ use WP_Statistics\Components\View;
             <?php endif; ?>
         </h2>
     <?php endif ?>
-
     <?php
     if (Menus::in_page('content-analytics') && Request::compare('type', 'single')) {
         View::load("components/headers/post-type");
@@ -39,6 +39,9 @@ use WP_Statistics\Components\View;
     }
     if (Menus::in_page('author-analytics') && Request::compare('type', 'single-author')) {
         View::load("components/headers/author-analytics");
+    }
+    if ((Menus::in_page('download_tracker') || Menus::in_page('link_tracker')) && Request::compare('type', 'single')) {
+        View::load("components/headers/tracker");
     }
     ?>
 
@@ -77,7 +80,7 @@ use WP_Statistics\Components\View;
     <?php endif ?>
 
     <?php if (isset($hasDateRang) || isset($filters) || isset($searchBoxTitle) || isset($filter)): ?>
-         <div class="<?php echo (Menus::in_page('content-analytics') || Menus::in_page('category-analytics') || Menus::in_page('author-analytics')) && (Request::compare('type', 'single') || Request::compare('type', 'single-author') )? 'wps-head-filters wps-head-filters--custom' : 'wps-head-filters' ?>">
+        <div class="<?php echo (Menus::in_page('content-analytics') || Menus::in_page('category-analytics') || Menus::in_page('author-analytics') || Menus::in_page('download_tracker') || Menus::in_page('link_tracker')) && (Request::compare('type', 'single') || Request::compare('type', 'single-author')) ? 'wps-head-filters wps-head-filters--custom' : 'wps-head-filters' ?>">
             <?php
             if (!empty($hasDateRang)) {
                 include 'date.range.php';