+++ includes/class-wp-statistics-db.php
@@ -29,7 +29,12 @@ class DB
         /**
          * Advanced Reporting Table
          */
-        'ar_outbox'
+        'ar_outbox',
+
+        /**
+         * Marketing Tables
+         */
+        'campaigns',
     );
 
     /**