+++ includes/class-wp-statistics-option.php
@@ -75,7 +75,9 @@ class Option
             'charts_previous_period'          => true,
             'attribution_model'               => 'first-touch',
             'geoip_location_detection_method' => 'maxmind',
-            'delete_data_on_uninstall'        => false
+            'delete_data_on_uninstall'        => false,
+            'share_anonymous_data'             => false,
+            'display_notifications'           => true,
         );
 
         return $options;