+++ includes/class-wp-statistics-visitor.php
@@ -7,6 +7,7 @@ use WP_Statistics\Service\Analytics\DeviceDetection\DeviceHelper;
 use WP_Statistics\Service\Analytics\VisitorProfile;
 use WP_Statistics\Service\Database\DatabaseFactory;
 use WP_Statistics\Service\Geolocation\GeolocationFactory;
+use WP_Statistics\Utils\Url;
 
 class Visitor
 {
@@ -164,15 +165,14 @@ class Visitor
             $visitor_id = $same_visitor->ID;
 
             // Update Same Visitor Hits
-            if ($args['exclusion_reason'] != 'Honeypot' and $args['exclusion_reason'] != 'Robot threshold') {
+            if ($args['exclusion_reason'] != 'Robot threshold') {
 
                 // Action Before Visitor Update
                 do_action('wp_statistics_update_visitor_hits', $visitor_id, $same_visitor);
 
                 $data = [
                     'hits'      => $same_visitor->hits + 1,
-                    'user_id'   => ! empty($same_visitor->user_id) ? $same_visitor->user_id : $visitorProfile->getUserId(),
-                    
+                    'user_id'   => ! empty($same_visitor->user_id) ? $same_visitor->user_id : $visitorProfile->getUserId()
                 ];
 
                 if (DatabaseFactory::compareCurrentVersion('14.12.6', '>=')) {
@@ -444,7 +444,7 @@ class Visitor
         global $wpdb;
 
         // Default Params
-        $params = array('link' => '', 'title' => '');
+        $params = array('link' => '', 'title' => '', 'query' => '');
 
         $pageTable = DB::table('pages');
 
@@ -454,13 +454,9 @@ class Visitor
             ARRAY_A);
 
         if ($item !== null) {
-            $params         = Pages::get_page_info($item['id'], $item['type'], $item['uri']);
-            $linkWithParams = !empty($item['uri']) ? home_url() . $item['uri'] : false;
-
-            // If URL has params, add it to the title (except for allowed params like UTM params, etc...)
-            if (trim($params['link'], '/') !== trim($linkWithParams, '/') && !Helper::checkUrlForParams($linkWithParams, Helper::get_query_params_allow_list())) {
-                $params['title'] .= ' (' . trim($item['uri']) . ')';
-            }
+            $params             = Pages::get_page_info($item['id'], $item['type'], $item['uri']);
+            $linkWithParams     = !empty($item['uri']) ? home_url() . $item['uri'] : '';
+            $params['query']    = Url::getParams($linkWithParams);
         }
 
         return $params;