+++ includes/class-wp-statistics.php
@@ -22,12 +22,14 @@ use WP_Statistics\Service\Admin\Posts\PostsManager;
 use WP_Statistics\Service\Admin\PrivacyAudit\PrivacyAuditManager;
 use WP_Statistics\Service\Admin\Referrals\ReferralsManager;
 use WP_Statistics\Service\Admin\TrackerDebugger\TrackerDebuggerManager;
+use WP_Statistics\Service\Admin\Notification\NotificationManager;
 use WP_Statistics\Service\Analytics\AnalyticsManager;
 use WP_Statistics\Service\Integrations\IntegrationsManager;
 use WP_Statistics\Service\Admin\FilterHandler\FilterManager;
 use WP_Statistics\Service\Admin\VisitorInsights\VisitorInsightsManager;
 use WP_Statistics\Service\Database\Managers\MigrationHandler;
 use WP_Statistics\Service\HooksManager;
+use WP_Statistics\Service\Admin\AnonymizedUsageData\AnonymizedUsageDataManager;
 
 defined('ABSPATH') || exit;
 
@@ -166,9 +168,11 @@ final class WP_Statistics
         // Ajax area
         require_once WP_STATISTICS_DIR . 'includes/admin/class-wp-statistics-admin-template.php';
 
-        $referrals      = new ReferralsManager();
-        $postsManager   = new PostsManager();
-        $userOnline     = new \WP_STATISTICS\UserOnline();
+        $referrals                  = new ReferralsManager();
+        $postsManager               = new PostsManager();
+        $userOnline                 = new \WP_STATISTICS\UserOnline();
+        $anonymizedUsageDataManager = new AnonymizedUsageDataManager();
+        $notificationManager        = new NotificationManager();
 
         // Admin classes
         if (is_admin()) {