+++ src/Components/Assets.php
@@ -29,6 +29,20 @@ class Assets
      */
     public static $plugin_dir = WP_STATISTICS_DIR;
 
+
+    /**
+     * Check if a script has been enqueued or not
+     *
+     * @param string $handle The script handle.
+     *
+     * @return bool
+     */
+    public static function isScriptEnqueued($handle)
+    {
+        $handle = self::getHandle($handle);
+        return wp_script_is($handle, 'enqueued');
+    }
+
     /**
      * Enqueue a script.
      *
@@ -52,11 +66,11 @@ class Assets
         $handle   = self::getHandle($handle);
         $version  = empty($version) ? WP_STATISTICS_VERSION : trim($version);
         $args     = $inFooter;
-        
+
         global $wp_version;
         $supportStrategy = version_compare($wp_version, '6.3', '>=');
 
-        if ($supportStrategy && ! empty($strategy)) {
+        if ($supportStrategy && !empty($strategy)) {
             $args = [
                 'in_footer' => $inFooter,
                 'strategy'  => $strategy,
@@ -64,12 +78,32 @@ class Assets
         }
 
         wp_enqueue_script($handle, self::getSrc($src, $obfuscate, $pluginUrl), $deps, $version, $args);
-        
+
         if ($localize) {
             $localize = apply_filters("wp_statistics_localize_{$handle}", $localize);
 
             wp_localize_script($handle, $object, $localize);
         }
+    }
+
+
+    /**
+     * Localize a script.
+     *
+     * @param string $handle The script handle.
+     * @param string $name The name of the object to be passed to the script.
+     * @param array $data An array of data to be localized.
+     *
+     * @return  void
+     * @example Assets::localize('admin', 'foo', ['bar' => 'baz']);
+     */
+    public static function localize($handle, $name, $data)
+    {
+        $handle = self::getHandle($handle);
+        $object = self::getObject($name);
+        $data   = apply_filters("wp_statistics_localize_{$handle}", $data);
+
+        wp_localize_script($handle, $object, $data);
     }
 
     /**
@@ -147,7 +181,7 @@ class Assets
         if ($obfuscate) {
             $file = $plugin_url ? Helper::urlToDir($plugin_url) : self::$plugin_dir;
             $file = new AssetNameObfuscator(path_join($file, self::$asset_dir . '/' . $src));
-            return $file->getHashedFileUrl();
+            return $file->getUrlThroughProxy();
         }
 
         $url = $plugin_url ? untrailingslashit($plugin_url) . '/' : self::$plugin_url;