+++ views/components/objects/external-link.php
@@ -1,1 +1,6 @@
-<a class="wps-link-arrow wps-link-arrow--external" title="<?php echo esc_html($title)?>" target="_blank"  href="<?php echo esc_url($url); ?>"><span><?php echo esc_html($title)?></span></a>
+<a class="wps-link-arrow wps-link-arrow--external"  target="_blank" href="<?php echo esc_url($url); ?>">
+    <span title="<?php echo esc_html($title) ?>"><?php echo esc_html($title) ?></span>
+    <?php if (!empty($tooltip)): ?>
+        <span class="wps-tooltip" title="<?php echo esc_attr($tooltip) ?>"><i class="wps-tooltip-icon info"></i></span>
+    <?php endif; ?>
+</a>