+++ views/pages/visitor-insights/single-visitor.php
@@ -11,7 +11,7 @@ $visitor = $data['visitor'];
     <div class="postbox-container postbox-container--first-col">
         <div class="wps-card">
             <div class="wps-card__title">
-                <h2><?php esc_html_e('Visitor Details', 'wp-statistics'); ?></h2>
+                <h2><?php esc_html_e('Session Details', 'wp-statistics'); ?></h2>
             </div>
             <?php View::load("components/session-details", ['visitor' => $visitor]); ?>
         </div>
@@ -39,24 +39,6 @@ $visitor = $data['visitor'];
                 View::load("components/tables/recent-views", $args);
                 ?>
             </div>
-
         </div>
-        <!-- <div class="wps-card wps-card--table">
-            <div class="wps-card__title">
-                <h2>
-                    <?php esc_html_e('Recent Events', 'wp-statistics'); ?>
-                    <span class="wps-tooltip" title="<?php esc_html_e('Recent Events tooltip', 'wp-statistics'); ?>"><i class="wps-tooltip-icon info"></i></span>
-                </h2>
-            </div>
-            <div class="wps-card--table__body">
-                <?php
-                $events_args = [
-                    'data'       => ['test', 'test'],
-                    'pagination' => isset($pagination) ? $pagination : null
-                ];
-                View::load("components/tables/recent-events", $events_args);
-                ?>
-            </div>
-        </div> -->
     </div>
 </div>